Cleaned up Details
